+++ app/Controllers/Opportunities/organization_controller.py
@@ -5,10 +5,10 @@ from uuid import UUID
 from sqlmodel import Session
 from Entities.OpportunityDTOs.organization_entity import CreateOrganization, UpdateOrganization, ReadOrganization
 from Services.Opportunities.organization_service import OrganizationService
-from Settings.logging_config import setup_logging
+from Settings.logging_config import get_logger
 from db import get_session
 
-logger = setup_logging()
+logger = get_logger()
 
 router = APIRouter(prefix="/Dijkstra/v1/organizations", tags=["Organizations"])
