+++ app/Controllers/Opportunities/projects_opportunities_controller.py
@@ -5,10 +5,10 @@ from uuid import UUID
 from sqlmodel import Session
 from Entities.OpportunityDTOs.projects_opportunities_entity import CreateProject, UpdateProject, ReadProject
 from Services.Opportunities.projects_opportunities_service import ProjectsOpportunitiesService
-from Settings.logging_config import setup_logging
+from Settings.logging_config import get_logger
 from db import get_session
 
-logger = setup_logging()
+logger = get_logger()
 
 router = APIRouter(prefix="/Dijkstra/v1/projects/opportunities", tags=["ProjectsOpportunities"])
