+++ app/Controllers/User/certifications_controller.py
@@ -3,11 +3,11 @@ from typing import List, Optional
 from uuid import UUID
 
 from sqlmodel import Session
-from Settings.logging_config import setup_logging
+from Settings.logging_config import get_logger
 from Services.User.certifications_service import CertificationService
 from Entities.UserDTOs.certification_entity import CreateCertification, UpdateCertification, ReadCertification
 from db import get_session
-logger = setup_logging()
+logger = get_logger()
 
 router = APIRouter(prefix="/Dijkstra/v1/certifications", tags=["Certifications"])
