+++ app/Controllers/User/dijkstra_certificate_controller.py
@@ -1,9 +1,9 @@
 from fastapi import APIRouter
-from Settings.logging_config import setup_logging
+from Settings.logging_config import get_logger
 from Services.User.dijkstra_certificate_service import CertificateGeneratorService
 
 # Initialize logging
-logger = setup_logging()
+logger = get_logger()
 
 router = APIRouter(prefix="/Dijkstra/v1/certificate", tags=["Certificate"])
