+++ app/Controllers/User/education_controller.py
@@ -5,10 +5,10 @@ from sqlmodel import Session
 
 from Entities.UserDTOs.education_entity import CreateEducation, UpdateEducation, ReadEducation
 from Services.User.education_service import EducationService
-from Settings.logging_config import setup_logging
+from Settings.logging_config import get_logger
 from db import get_session
 
-logger = setup_logging()
+logger = get_logger()
 
 router = APIRouter(prefix="/Dijkstra/v1/education", tags=["Education"])
