+++ app/Controllers/User/location_controller.py
@@ -5,10 +5,10 @@ from sqlmodel import Session
 
 from Entities.UserDTOs.location_entity import CreateLocation, UpdateLocation, ReadLocation
 from Services.User.location_service import LocationService
-from Settings.logging_config import setup_logging
+from Settings.logging_config import get_logger
 from db import get_session
 
-logger = setup_logging()
+logger = get_logger()
 
 router = APIRouter(prefix="/Dijkstra/v1/location", tags=["Locations"])
