+++ app/Controllers/User/projects_controller.py
@@ -3,7 +3,7 @@ from uuid import UUID
 from fastapi import APIRouter, Depends
 from sqlmodel import Session
 
-from Settings.logging_config import setup_logging
+from Settings.logging_config import get_logger
 from Entities.UserDTOs.projects_entity import (
     CreateProject,
     DeleteResponse,
@@ -13,7 +13,7 @@ from Entities.UserDTOs.projects_entity import (
 from Services.User.projects_service import ProjectsService
 from db import get_session
 
-logger = setup_logging()
+logger = get_logger()
 
 router = APIRouter(prefix="/Dijkstra/v1/projects", tags=["Projects"])
