+++ app/Controllers/User/test_scores_controller.py
@@ -3,11 +3,11 @@ from typing import List, Optional
 from uuid import UUID
 
 from sqlmodel import Session
-from Settings.logging_config import setup_logging
+from Settings.logging_config import get_logger
 from Services.User.test_scores_service import TestScoresService
 from Entities.UserDTOs.test_scores_entity import CreateTestScore, UpdateTestScore, ReadTestScore
 from db import get_session
-logger = setup_logging()
+logger = get_logger()
 
 router = APIRouter(prefix="/Dijkstra/v1/test-scores", tags=["Test Scores"])
