+++ app/Services/User/certifications_service.py
@@ -1,4 +1,4 @@
-from Settings.logging_config import setup_logging
+from Settings.logging_config import get_logger
 from uuid import UUID
 from typing import List, Optional
 from sqlmodel import Session
@@ -8,7 +8,7 @@ from Utils.Exceptions.user_exceptions import CertificationNotFound
 from Entities.UserDTOs.certification_entity import CreateCertification, UpdateCertification
 
 
-logger = setup_logging()
+logger = get_logger()
 
 class CertificationService:
