+++ app/main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from Settings.logging_config import setup_logging
+from Settings.logging_config import get_logger
 from Controllers import main_controller
 from Controllers.Opportunities import job_controller
 from Controllers.User import (
@@ -8,7 +8,6 @@ from Controllers.User import (
     dijkstra_certificate_controller,
     document_controller,
     education_controller,
-    leetcode_controller,
     links_controller,
     profile_controller,
     projects_controller,
@@ -42,7 +41,7 @@ app.add_middleware(
 )
 
 # Initialize logging
-logger = setup_logging()
+logger = get_logger()
 
 @app.on_event("startup")
 def on_startup():
@@ -61,7 +60,6 @@ app.include_router(user_controller.router)
 app.include_router(workexperience_controller.router)
 app.include_router(location_controller.router)
 app.include_router(profile_controller.router)
-app.include_router(leetcode_controller.router)
 app.include_router(dijkstra_certificate_controller.router)
 app.include_router(certifications_controller.router)
 app.include_router(document_controller.router)
